backend2/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
PyMySQL==1.1.0
flasgger==0.9.7.1
marshmallow==3.20.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0